feat(python): add ptr for small integer types
