+++ py-polars/src/series/mod.rs
@@ -671,43 +671,6 @@ impl PySeries {
         let out = hist(&self.series, bins.as_ref(), bin_count).map_err(PyPolarsErr::from)?;
         Ok(out.into())
     }
-
-    fn get_ptr(&self) -> PyResult<usize> {
-        let s = self.series.to_physical_repr();
-        let arrays = s.chunks();
-        if arrays.len() != 1 {
-            let msg = "Only can take pointer, if the 'series' contains a single chunk";
-            raise_err!(msg, ComputeError);
-        }
-        match s.dtype() {
-            DataType::Boolean => {
-                let ca = s.bool().unwrap();
-                let arr = ca.downcast_iter().next().unwrap();
-                // this one is quite useless as you need to know the offset
-                // into the first byte.
-                let (slice, start, _len) = arr.values().as_slice();
-                if start == 0 {
-                    Ok(slice.as_ptr() as usize)
-                } else {
-                    let msg = "Cannot take pointer boolean buffer as it is not perfectly aligned.";
-                    raise_err!(msg, ComputeError);
-                }
-            }
-            dt if dt.is_numeric() => Ok(with_match_physical_numeric_polars_type!(s.dtype(), |$T| {
-                let ca: &ChunkedArray<$T> = s.as_ref().as_ref().as_ref();
-                get_ptr(ca)
-            })),
-            _ => {
-                let msg = "Cannot take pointer of nested type";
-                raise_err!(msg, ComputeError);
-            }
-        }
-    }
-}
-
-fn get_ptr<T: PolarsNumericType>(ca: &ChunkedArray<T>) -> usize {
-    let arr = ca.downcast_iter().next().unwrap();
-    arr.values().as_ptr() as usize
 }
 
 macro_rules! impl_set_with_mask {
